@@ -1,26 +1,25 @@
 name: cpx_research_sdk_flutter
 description: CPX Research SDK Flutter
-version: 0.3.1
-author: Dualmeta
+version: 0.4.1
 homepage: https://www.dualmeta.io
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
   http: ^0.13.3
-  webview_flutter: ^2.0.9
-  json_annotation: ^3.0.1
+  webview_flutter: ^3.0.1
+  json_annotation: ^4.4.0
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.7.3
-  json_serializable: ^3.2.5
+  build_runner: ^2.1.2
+  json_serializable: ^6.1.4
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
